+++ client/src/components/utilities/EC2InstanceDetail.tsx
