+++ client/src/components/utilities/SignUp.tsx
@@ -1,4 +1,3 @@
-// /src/features/auth/SignUp.tsx
 import React from 'react';
 
 const SignUp: React.FC = () => {
