--- client/src/features/auth/Login.tsx
@@ -1,21 +1,0 @@
-// /src/features/auth/Login.tsx
-import React from 'react';
-import { useDispatch } from 'react-redux';
-import { login } from './authSlice';
-
-const Login: React.FC = () => {
-  const dispatch = useDispatch();
-
-  const handleLogin = () => {
-    dispatch(login());
-  };
-
-  return (
-    <div>
-      <h1>Login</h1>
-      <button onClick={handleLogin}>Login</button>
-    </div>
-  );
-};
-
-export default Login;
