--- client/src/features/common/Footer.tsx
@@ -1,12 +1,0 @@
-// /src/features/common/Footer.tsx
-import React from 'react';
-
-const Footer: React.FC = () => {
-  return (
-    <footer>
-      <p>© 2024 My App</p>
-    </footer>
-  );
-};
-
-export default Footer;
