--- client/src/features/common/Header.tsx
@@ -1,12 +1,0 @@
-// /src/features/common/Header.tsx
-import React from 'react';
-
-const Header: React.FC = () => {
-  return (
-    <header>
-      <h1>My App</h1>
-    </header>
-  );
-};
-
-export default Header;
