+++ client/src/pages/LoginPage.tsx
@@ -1,13 +1,8 @@
-// /src/components/Input.tsx
 import React from 'react';
-import Login from '../features/auth/Login';
+// import Login from '../components/auth/Login';
 
 const LoginPage: React.FC = () => {
-  return (
-    <div>
-      <Login />
-    </div>
-  );
+  return <div>{/* <Login /> */}</div>;
 };
 
 export default LoginPage;
