+++ client/src/store.ts
@@ -1,18 +1,5 @@
-// /src/app/store.ts
-// import { configureStore } from '@reduxjs/toolkit';
-// import rootReducer from './rootReducer';
-
-// const store = configureStore({
-//   reducer: rootReducer,
-// });
-
-// export type AppDispatch = typeof store.dispatch;
-// export type RootState = ReturnType<typeof store.getState>;
-
-// export default store;
-// store.ts
 import { configureStore, ThunkAction, Action } from '@reduxjs/toolkit';
-import rootReducer from './rootReducer';
+import rootReducer from './reducers/rootReducer';
 
 export const store = configureStore({
   reducer: rootReducer,
