+++ package.json
@@ -5,7 +5,6 @@
   "main": "",
   "scripts": {
     "clean": "rm -rf ./dist/*",
-    "build": "npm start",
     "dev": "webpack serve --config webpack.config.ts --open",
     "start": "npm run clean && tsc && nodemon dist/server/src/server.js",
     "test": "echo \"Error: no test specified\" && exit 1",
@@ -15,7 +14,7 @@
     "type": "git",
     "url": "git+https://github.com/oslabs-beta/SkyScraper.git"
   },
-  "author": "Tripp Murphy, Christie Laferriere, Bin He, Abel, Nikola",
+  "author": "Tripp Murphy, Christie Laferriere, Bin He, Abel Ratanaphan, Nikola Andelkovic",
   "license": "MIT",
   "bugs": {
     "url": "https://github.com/oslabs-beta/SkyScraper/issues"
