+++ tsconfig.json
@@ -2,28 +2,32 @@
   "compilerOptions": {
     "allowJs": true,
     "target": "ES6", //"target": "es6", // Target ECMAScript version
-    //"allowImportingTsExtensions": true, // Or a specific ECMAScript version that supports ES Modules
     "noEmit": false, // Added noEmit to complement allowImportingTsExtensions
     "module": "CommonJS", //"module": "esnext", // Enable ES module syntax
     "moduleResolution": "node", // Use Node.js module resolution
-    // "jsx": "react-jsx", // Enable JSX for React
     "jsx": "react", // Enable JSX for React
     "strict": true, // Enable all strict type-checking options
     "esModuleInterop": true, // Enables default imports from modules with no default export
     "outDir": "./dist",
     "skipLibCheck": true, // Skip type checking of declaration files
     "resolveJsonModule": true,
+    "baseUrl": "./", // Base directory to resolve non-relative module names
+    // "jsx": "react-jsx", // Enable JSX for React
+    //"allowImportingTsExtensions": true, // Or a specific ECMAScript version that supports ES Modules
     //"forceConsistentCasingInFileNames": true, // Disallow inconsistently-cased references to the same file
     //"baseUrl": ".", // Base directory to resolve non-relative module names
-    "baseUrl": "./", // Base directory to resolve non-relative module names
     "paths": {
       // "*": ["node_modules/*", "/server/src/*"] // Path mapping for custom type definitions
       "*": ["node_modules/*", "src/*"]
     },
     "lib": ["dom", "dom.iterable", "esnext"] // Library files to be included in the compilation
   },
-  "include": ["client/src/**/*", "server/src/**/*", "webpack.config.ts"], // Include all TypeScript files in the src directory
-  // "exclude": ["node_modules", "**/*.spec.ts"] // Exclude node_modules and test files
+  "include": [
+    "client/src/**/*",
+    "server/src/**/*",
+    "webpack.config.ts",
+    "client/src/components/auth"
+  ], // Include all TypeScript files in the src directory
   "exclude": ["node_modules", "**/*.spec.js"] // Exclude node_modules and test files
   //-------
   // "forceConsistentCasingInFileNames": true // Disallow inconsistently-cased references to the same file.
